+++ packages/toolkit/src/view/settings/integrations/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Integrations";
