--- packages/toolkit/src/view/settings/user/UserIntegrationsTab/index.tsx
@@ -1,1 +1,0 @@
-export * from "./UserIntegrationsTab";
